Feat/orders (#27)
* fixes

* many updates for orders and updates for structure of other files

* wip: gili

* navigate to my orders from navbar

* navigate admin orders works

* mark order as arrived

* added photoes to workouts

* changed arrived to confirmed

* minor changes to orders confirmed

---------

+++ public/js/cart.js
@@ -33,6 +33,17 @@ async function redirectToCart() {
 
 async function confirmOrder() {
     try {
+        const createOrder = await fetch('/orders', {
+            method: 'POST',
+            headers: {
+                'Content-Type': 'application/json'
+            },
+        });
+
+        if (!createOrder.ok) {
+            console.error('Failed to create order');
+            return
+        }
         // Send API request to remove workout from cart
         const deleteResponse = await fetch('/cart/deleteAllWorkoutsFromCart', {
             method: 'POST',
